Revert "Lauren (#18)"
This reverts commit 08110bd5d673341413a38118c395167deceafcd7.
