--- .env.development
@@ -1,1 +1,0 @@
-HOST=https://waste-not-smu.herokuapp.com
