+++ .gitignore
@@ -1,22 +1,19 @@
-# See https://help.github.com/ignore-files/ for more about ignoring files.
-
 # dependencies
+.env
 /node_modules
+/config/keys.js
 
 # testing
 /coverage
 
 # production
 /build
-/config
 /sessions
-./config
+./config/keys.js
 /sessions
-
 
 # misc
 .DS_Store
-.env
 .env.local
 .env.development.local
 .env.test.local
@@ -25,10 +22,3 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-
-
-waste-not-firebase-adminsdk-dy000-2aff960931.json
-node_modules
-node_modules
-node_modules
-npm-debug.log
