+++ client/.gitignore
@@ -19,5 +19,3 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-node_modules
-npm-debug.log
