+++ client/public/index.html
@@ -26,15 +26,15 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>Love Food Not Waste</title>
-</head>
+        <title>React App</title>
+    </head>
 
-<body className='Site'>
-    <noscript>
-        You need to enable JavaScript to run this app.
-    </noscript>
-    <div id="root"></div>
-    <!--
+    <body>
+        <noscript>
+            You need to enable JavaScript to run this app.
+        </noscript>
+        <div id='root'> </div>
+        <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
