+++ client/src/components/Footer/Footer.js
@@ -4,7 +4,7 @@ import "./Footer.css";
 const Footer = () => (
 
     <footer>
-        <p className='text-center' id='footerText'> &copy; 2018 Built by Lauren Abney, Avery Bruner, Jade Gonzales &amp; Courtney Montgomery</p>
+        <p className='text-center' id='footerText'> &copy; 2018 Built by Lauren Abney,  Avery Bruner, Jade Gonzales &amp; Courtney Montgomery</p>
     </footer>
 
 );
