+++ client/src/pages/AtHome/Recipes.css
@@ -1,13 +1,8 @@
-
-
 .recipeImg {
     border-radius: 2px;
     margin: 0 auto;
   
 }
-
-
-
 .recipes {
     width: 40%;
     margin: 0 auto;
