+++ client/src/pages/AtHome/index.js
@@ -1,1 +1,2 @@
+export * from './Recipes';
 export * from './Inventory';
