--- client/src/pages/Chat.css
@@ -1,107 +1,0 @@
-
-
-.App {
-  background-color: red;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.chatroom {
-  width: 360px;
-  height: 425px;
-  color: white;
-  font-size: 13px;
-  text-align: center;
-  background-color: rgba(0,0,0, .7);
-  border-radius: 5px;
-  overflow: hidden;
-  box-shadow: 0 0 8px 0 rgba(0,0,0, 0.3);
-  position: absolute;
-    left: 37%;
-    
-}
-
-.chatroom h3 {
-  background-color: red;
-  color: #FFF;
-  margin: 0;
-  height: 40px;
-  font-size: 15px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.chatroom .chats {
-  box-sizing: border-box;
-  padding: 0 20px;
-  margin: 10px 0 0;
-  height: 335px;
-  overflow-y: scroll;
-}
-
-.chatroom .chats .chat {
-  background: red;
-  position: relative;
-  padding: 5px 13px;
-  font-size: 14px;
-  border-radius: 10px;
-  list-style: none;
-  float: left;
-  clear: both;
-  margin: 10px 0;
-  max-width: 200px;
-}
-
-.chatroom .chats .chat.right {
-  float: right;
-  clear: both;
-}
-
-.chatroom .chats .chat a {
-  text-decoration: none;
-  color: red;
-}
-
-.chatroom .input {
-  height: 40px;
-  width: 100%;
-  background: red;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.chatroom .input input[type="text"] {
-  background: rgba(255, 255, 255, 0.5);
-  outline: 0;
-  border: 2px solid red;
-  padding: 0 5px;
-  border-top-left-radius: 5px;
-  border-bottom-left-radius: 5px;
-  margin-left: 10px;
-  height: 30px;
-  width: 270px;
-  box-sizing: border-box;
-  font-size: 14px;
-}
-
-.chatroom .input input[type="submit"] {
-  background: transparent;
-  border: 1px solid red;
-  height: 30px;
-  width: 70px;
-  font-size: 15px;
-  color: white;
-  box-sizing: border-box;
-  border-top-right-radius: 5px;
-  border-bottom-right-radius: 5px;
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-.chatroom .input input[type="submit"]:hover {
-  background: red;
-  color: #FFF;
-}
