--- client/src/pages/Recipes/index.js
@@ -1,2 +1,0 @@
-export * from './Recipes';
-export * from './FullRecipe';
