--- controllers/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-    inventoryController: require( './inventoryController' ),
-    foodController:  require('./foodController')
-};
