--- models/index.js
@@ -1,5 +1,0 @@
-// Export Database Schemas to be used throughout project
-module.exports = {
-    Inventory: require( './inventory' ),
-    Food:  require('./food')
-};
