--- routes/index.js
@@ -1,14 +1,0 @@
-const path = require( 'path' );
-const router = require( 'express' ).Router();
-const apiRoutes = require( './api' );
-// const userRoutes = require( './auth/userRoutes' );
-
-// API Routes
-router.use( '/api', apiRoutes );
-
-// If no API routes are hit, send the React app
-router.use( function ( req, res ) {
-    res.sendFile( path.join( __dirname, '../client/public/index.html' ) );
-} );
-
-module.exports = router;
