--- scripts/build.js
@@ -1,3 +1,0 @@
-const args = ["run build"];
-const opts = { stdio: "inherit", cwd: "client", shell: true };
-require("child_process").spawn("npm", args, opts);
