--- seeds/food.json
@@ -1,6 +1,0 @@
-[
-    { "category":"Fruit", "item": "Apples, whole","pantry":"2-4 Weeks","refrigerator":"1-2 Months"},
-    { "category":"Fruit", "item": "Apples, fresh cut","pantry":"","refrigerator":"3 to 5 Days"},
-    { "category":"Fruit", "item": "Applesauce","pantry":"1-2 Months (unopened)","refrigerator":"1-2 Weeks (opened)"},
-    { "category":"Dessert","item":"Apple Pie","pantry":"1-2 Days","refrigerator":"4-5 Days"}
-]
