--- views/views.txt
@@ -1,1 +1,0 @@
-Directory to store any views or templates we may be using (i.e. Handlebars)
